requirements.txt
beautifulsoup4==4.6.3
boilerpipe3==1.1
charade==1.0.3
chardet==3.0.4
JPype1==0.6.3
JPype1-py3==0.5.5.2
lxml==4.2.4
mp-throttle==0.0.5
nltk==3.3
numpy==1.15.1
pdfminer.six==20170720
pkg-resources==0.0.0
pycryptodome==3.6.6
PyYAML==3.13
scikit-learn==0.19.2
scipy==1.1.0
six==1.11.0
sklearn==0.0
SQLAlchemy==1.2.11
